+++ tests/testthat/test-df2rast.R
@@ -6,4 +6,6 @@ test_that("check if the object class is correct", {
   # tests
   expect_s4_class(phyloraster::df2rast(dat$presab),
                   "SpatRaster")
+  expect_s4_class(phyloraster::df2rast(dat$presab, CRS = "EPSG:2169"),
+                  "SpatRaster")
 })