+++ tests/testthat/test-range-size.R
@@ -6,8 +6,10 @@ test_that("Are the returned values correct", {
   x <- x[[1:5]]
 
   rs.obs <- range_size(x, terra::cellSize(x))
+  rs.obs2 <- range_size(x)
   rs.expec <- c(5021016056, 196704341368, 4523101328,
                 26914371304, 216777518096)
 
   expect_equivalent(rs.obs, rs.expec)
+  expect_equivalent(rs.obs2, rs.expec)
 })