+++ NEWS.md
@@ -1,3 +1,12 @@
+# phyloraster 2.1.0
+
+# phyloraster 2.1.0
+
+- added functions to compute full and tip branch lengths: 'species.tip.length()' and 'species.branch.length()'
+- added new function to compute ED for each species: 'species.ed()'
+- fixed ED calculation for raster data
+- added new tests for the new functions
+
 # phyloraster 2.0.1
 
 * Initial CRAN submission.